refactor(engine): removing another public api isAttributeLocked
fix(engine): review feedback

fix(engine): review feedback

Update packages/@lwc/engine-core/src/framework/base-bridge-element.ts
