+++ packages/@lwc/engine-core/src/framework/base-bridge-element.ts
@@ -19,12 +19,15 @@ import {
     isUndefined,
     seal,
     setPrototypeOf,
+    keys,
+    htmlPropertyToAttribute
 } from '@lwc/shared';
 
 import { getAssociatedVM } from './vm';
 import { reactiveMembrane } from './membrane';
 import { HTMLElementConstructor } from './renderer';
 import { HTMLElementOriginalDescriptors } from './html-properties';
+import { isAttributeLocked } from './attributes';
 
 // A bridge descriptor is a descriptor whose job is just to get the component instance
 // from the element instance, and get the value or set a new value on the component.
@@ -67,6 +70,50 @@ function createMethodCaller(methodName: string): (...args: any[]) => any {
     };
 }
 
+type AttributeChangedCallback = (
+    this: HTMLElement,
+    attrName: string,
+    oldValue: string,
+    newValue: string
+) => void;
+
+function createAttributeChangedCallback(
+    attributeToPropMap: Record<string, string>,
+    superAttributeChangedCallback?: AttributeChangedCallback
+): AttributeChangedCallback {
+    return function attributeChangedCallback(
+        this: HTMLElement,
+        attrName: string,
+        oldValue: string,
+        newValue: string
+    ) {
+        if (oldValue === newValue) {
+            // Ignore same values.
+            return;
+        }
+        const propName = attributeToPropMap[attrName];
+        if (isUndefined(propName)) {
+            if (!isUndefined(superAttributeChangedCallback)) {
+                // delegate unknown attributes to the super.
+                // Typescript does not like it when you treat the `arguments` object as an array
+                // @ts-ignore type-mismatch
+                superAttributeChangedCallback.apply(this, arguments);
+            }
+            return;
+        }
+        if (!isAttributeLocked(this, attrName)) {
+            // Ignore changes triggered by the engine itself during:
+            // * diffing when public props are attempting to reflect to the DOM
+            // * component via `this.setAttribute()`, should never update the prop
+            // Both cases, the setAttribute call is always wrapped by the unlocking of the
+            // attribute to be changed
+            return;
+        }
+        // Reflect attribute change to the corresponding property when changed from outside.
+        (this as any)[propName] = newValue;
+    };
+}
+
 export interface HTMLElementConstructor {
     prototype: HTMLElement;
     new (): HTMLElement;
@@ -104,10 +151,16 @@ export function HTMLBridgeElementFactory(
             value: HTMLBridgeElement,
         });
     }
+    // generating the hash table for attributes to avoid duplicate fields and facilitate validation
+    // and false positives in case of inheritance.
+    const attributeToPropMap: Record<string, string> = create(null);
+    const { attributeChangedCallback: superAttributeChangedCallback } = SuperClass.prototype as any;
+    const { observedAttributes: superObservedAttributes = [] } = SuperClass as any;
     const descriptors: PropertyDescriptorMap = create(null);
     // expose getters and setters for each public props on the new Element Bridge
     for (let i = 0, len = props.length; i < len; i += 1) {
         const propName = props[i];
+        attributeToPropMap[htmlPropertyToAttribute(propName)] = propName;
         descriptors[propName] = {
             get: createGetter(propName),
             set: createSetter(propName),
@@ -124,6 +177,20 @@ export function HTMLBridgeElementFactory(
             configurable: true,
         };
     }
+    // creating a new attributeChangedCallback per bridge because they are bound to the corresponding
+    // map of attributes to props. We do this after all other props and methods to avoid the possibility
+    // of getting overrule by a class declaration in user-land, and we make it non-writable, non-configurable
+    // to preserve this definition.
+    descriptors.attributeChangedCallback = {
+        value: createAttributeChangedCallback(attributeToPropMap, superAttributeChangedCallback),
+    };
+    // Specify attributes for which we want to reflect changes back to their corresponding
+    // properties via attributeChangedCallback.
+    defineProperty(HTMLBridgeElement, 'observedAttributes', {
+        get() {
+            return [...superObservedAttributes, ...keys(attributeToPropMap)];
+        },
+    });
     defineProperties(HTMLBridgeElement.prototype, descriptors);
     return HTMLBridgeElement as HTMLElementConstructor;
 }
